Refactor computer renaming logic and add support for custom prefix
This commit refactors the computer renaming logic in the RenameComputer.ps1 script. It introduces support for specifying a custom prefix for the computer name. If a prefix is provided, the new computer name will be in the format "Prefix-SerialNumber". If no prefix is provided, the default format "UQ-SerialNumber" will be used. Additionally, a bug in PowerShell 5.1 related to retrieving the computer serial number has been fixed.

Closes #<issue_number>